Reduce key timeout to 1
I've been having problems, particularly on some machines and with
longer-running instances of zsh, where hitting ESC and subsequently
/ to activate search mode doesn't work properly unless I add a pretty
substantial delay between the two keys.  This is annoying, and I found
this post that suggested lowering KEYTIMEOUT:

    https://dougblack.io/words/zsh-vi-mode.html

I'm concerned with how this will affect other keybindings, but let's
give it a whirl and see how it works.