Work around hardcoded vi search keymap behavior
When you enter search mode using vi-history-search-backward or vi-history-search-forward,
zsh, as of version 5.6.2, loads the main keymap and dispatches on
command values, rather than executing the widgets bound to the keys
themselves.  The dispatch table is limited to a very small number of
commands, and since it's coded into the C implementation, it obviously
doesn't include custom widgets such as pacsearch_replace.  If a command
value zsh doesn't recognize is encountered, it's ignored - which caused
my search usage not to allow typing since I wrote the pacsearch_replace
thing.  This fix works around that behavior