Add tar function
I pipe to tar all the time, and I hate how GNU tar doesn't detect the
filename based on the magic signature of the file.

(Actually, it _does_ - it just tells you that your flag is incompatible
and exits with an error)

So let's detect if we're being piped to - if so, use bsdtar, which
_does_ do this detection, otherwise use GNU tar, which is what I've been
using for years

Hopefully this doesn't break anything.