Fix sqlite3 completion bug creating ':memory:' files
TIL `words` contains the words as they appear on the command line,
verbatim - so quotes and all.  So if you have this:

  $ sqlite3 ':memory:' …

…because it's forever late at night and you forget if `:` is significant
to the shell or not in that context and you don't care you just want to
go to bed and a few extra quotes never hurt anyone and *then* you
tab-complete, `words` looks like this:

  words=(sqlite3 "':memory:'")

…which then ends up creating a SQLite URL that looks like this:

  file:':memory:'?immutable=1

…which `sqlite3` happily accepts and says "hey, `':memory:'` isn't my
special `:memory:` value - I'd better treat that as a filename!" and
*also* says "hey, that file doesn't exist - I'd better create it!",
which all leaves poor, tired Rob scratching his head like Kosuke
Kindaichi trying to figure out which rogue process is leaving all of
these cursed `':memory:'` files around.

Aaaaanyway, this change just tells zsh to strip any quotes around the
database filename, which allows `sqlite3` to get the special `:memory:`
value and do the right thing.